+++ apps/forms/components/pdf-viewer/index.ts
@@ -1,0 +1,1 @@
+export * from "./pdf-viewer";
