+++ apps/forms/scaffolds/blocks-editor/blocks/pdf-block.tsx
@@ -2,13 +2,9 @@
 
 import { useCallback, useState } from "react";
 import {
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  CodeIcon,
   DotsHorizontalIcon,
   ReaderIcon,
   TrashIcon,
-  VideoIcon,
 } from "@radix-ui/react-icons";
 import {
   DropdownMenu,
@@ -19,13 +15,7 @@ import {
 import { EditorFlatFormBlock } from "@/scaffolds/editor/state";
 import { BlockHeader, FlatBlockBase, useDeleteBlock } from "./base-block";
 import { useEditorState } from "@/scaffolds/editor";
-import { pdfjs } from "react-pdf";
-import { Document, Page } from "react-pdf";
-import "react-pdf/dist/Page/AnnotationLayer.css";
-import "react-pdf/dist/Page/TextLayer.css";
-
-// pdfjs worker
-pdfjs.GlobalWorkerOptions.workerSrc = `//unpkg.com/pdfjs-dist@${pdfjs.version}/build/pdf.worker.min.js`;
+import { PDFViewer } from "@/components/pdf-viewer";
 
 export function PdfBlock({
   id,
@@ -87,48 +77,5 @@ export function PdfBlock({
         </div>
       </div>
     </FlatBlockBase>
-  );
-}
-
-function PDFViewer({ file }: { file: string }) {
-  const [numPages, setNumPages] = useState<number>();
-  const [pageNumber, setPageNumber] = useState<number>(1);
-
-  function onDocumentLoadSuccess({ numPages }: { numPages: number }): void {
-    setNumPages(numPages);
-  }
-
-  return (
-    <div className="relative w-full h-full">
-      <Document
-        className="w-full h-full overflow-scroll"
-        file={file ?? ""}
-        onLoadSuccess={onDocumentLoadSuccess}
-      >
-        <Page pageNumber={pageNumber} />
-      </Document>
-      <footer
-        className="absolute bottom-0 left-0 right-0 bg-white dark:bg-gray-800 p-2.5 text-sm text-gray-700 dark:text-gray-300"
-        style={{ zIndex: 999 }}
-      >
-        <div className="flex gap-2">
-          <button
-            disabled={pageNumber <= 1}
-            onClick={() => setPageNumber((prev) => prev - 1)}
-          >
-            <ArrowLeftIcon />
-          </button>
-          <p>
-            Page {pageNumber} of {numPages}
-          </p>
-          <button
-            disabled={pageNumber >= (numPages ?? 1)}
-            onClick={() => setPageNumber((prev) => prev + 1)}
-          >
-            <ArrowRightIcon />
-          </button>
-        </div>
-      </footer>
-    </div>
   );
 }
